Modified e2e chain model for librispeech to work.

+++ egs/wsj/s5/steps/nnet3/chain/e2e/train_e2e.py
@@ -317,8 +317,8 @@ def train(args, run_opts):
                     {dir}/init.raw""".format(command=run_opts.command,
                                              dir=args.dir))
 
-    egs_left_context = left_context + args.frame_subsampling_factor / 2
-    egs_right_context = right_context + args.frame_subsampling_factor / 2
+    egs_left_context = int(left_context + args.frame_subsampling_factor / 2)
+    egs_right_context = int(right_context + args.frame_subsampling_factor / 2)
     egs_left_context_initial = (left_context_initial + args.frame_subsampling_factor / 2 if
                                 left_context_initial >= 0 else -1)
     egs_right_context_final = (right_context_final + args.frame_subsampling_factor / 2 if
@@ -413,7 +413,7 @@ def train(args, run_opts):
     logger.info("Training will run for {0} epochs = "
                 "{1} iterations".format(args.num_epochs, num_iters))
 
-    for iter in range(num_iters):
+    for iter in range(int(num_iters)):
 
         percent = num_archives_processed * 100.0 / num_archives_to_process
         epoch = (num_archives_processed * args.num_epochs
@@ -423,6 +423,9 @@ def train(args, run_opts):
             logger.info("Exiting early due to --exit-stage {0}".format(iter))
             return
 
+        # logger.info("args.num_jobs_initial: {0}".format(args.num_jobs_initial))
+        # logger.info("args.num_jobs_final: {0}".format(args.num_jobs_final))
+        # logger.info("args.num_jobs_step: {0}".format(args.num_jobs_step))
         current_num_jobs = common_train_lib.get_current_num_jobs(
             iter, num_iters,
             args.num_jobs_initial, args.num_jobs_step, args.num_jobs_final)
@@ -453,7 +456,7 @@ def train(args, run_opts):
                 shrink_info_str = 'shrink: {0:0.5f}'.format(shrinkage_value)
             logger.info("Iter: {0}/{1}   Jobs: {2}   "
                         "Epoch: {3:0.2f}/{4:0.1f} ({5:0.1f}% complete)   "
-                        "lr: {6:0.6f}   {7}".format(iter, num_iters - 1,
+                        "lr: {6:0.6f}   {7}".format(iter, int(num_iters - 1),
                                                     current_num_jobs,
                                                     epoch, args.num_epochs,
                                                     percent,
